Adds abstract speech visitor.
