+++ src/main/java/com/progressiveaccess/cmlspeech/sre/EnglishExpertSpeechVisitor.java
@@ -46,9 +46,6 @@ import com.progressiveaccess.cmlspeech.structure.RichMolecule;
 import com.progressiveaccess.cmlspeech.structure.RichSetType;
 import com.progressiveaccess.cmlspeech.structure.RichSubRing;
 
-import com.google.common.base.Joiner;
-
-import java.util.LinkedList;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
@@ -56,16 +53,9 @@ import java.util.TreeSet;
  * Produces the basic speech for structures.
  */
 
-public class EnglishExpertSpeechVisitor implements SpeechVisitor {
+public class EnglishExpertSpeechVisitor extends AbstractSpeechVisitor {
 
-  private ComponentsPositions contextPositions = null;
-  private LinkedList<String> speech = new LinkedList<String>();
   private boolean shortDescription = false;
-
-
-  public void setContextPositions(final ComponentsPositions positions) {
-    this.contextPositions = positions;
-  }
 
 
   @Override
@@ -76,7 +66,7 @@ public class EnglishExpertSpeechVisitor implements SpeechVisitor {
 
   @Override
   public void visit(final RichAtom atom) {
-    Integer position = this.contextPositions.getPosition(atom.getId());
+    Integer position = this.getContextPositions().getPosition(atom.getId());
     // TODO (sorge) Maybe take the supersystem of the atom outside the context.
     if (position == null) {
       this.describeSuperSystem(atom);
@@ -210,24 +200,6 @@ public class EnglishExpertSpeechVisitor implements SpeechVisitor {
   }
 
 
-  private void modSpeech(final String msg) {
-    String last = this.speech.removeLast();
-    this.speech.offerLast(last + msg);
-  }
-
-
-  private void addSpeech(final String msg) {
-    if (!msg.equals("")) {
-      this.speech.add(msg);
-    }
-  }
-
-
-  private void addSpeech(final Integer num) {
-    this.addSpeech(num.toString());
-  }
-
-
   // TODO (sorge) Do something about all upper case names without destroying
   // important upper cases. E.g.: WordUtils.capitalizeFully.
   private void addName(final RichAtomSet atomset) {
@@ -240,14 +212,6 @@ public class EnglishExpertSpeechVisitor implements SpeechVisitor {
       return;
     }
     addSpeech(atomset.getMolecularFormula());
-  }
-
-
-  public String getSpeech() {
-    final Joiner joiner = Joiner.on(" ");
-    String result = joiner.join(this.speech);
-    this.speech.clear();
-    return result + ".";
   }
 
 
@@ -271,7 +235,7 @@ public class EnglishExpertSpeechVisitor implements SpeechVisitor {
           this.addSpeech(position);
           this.addSpeech("and");
         }
-        this.speech.removeLast();
+        this.remSpeech();
     }
   }
